% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.panglm}
\alias{fitted.panglm}
\title{Extract fitted values from a panglm model}
\usage{
\method{fitted}{panglm}(object, ...)
}
\arguments{
\item{object}{a fitted `"panglm"` object}

\item{...}{unused}
}
\description{
Values are returned in the row order of the data supplied to [panglm()],
even though estimation internally sorts observations by panel index.
Fixed-effects binomial values are conditional inclusion probabilities,
conditional on each individual's observed number of successes.
}
