% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panglm-package.R
\docType{package}
\name{panglm-package}
\alias{panglm-package}
\title{panglm: Generalized Linear Models for Panel Data}
\description{
Provides generalized linear models for panel data, including pooled, fixed-effects, and random-effects estimators for continuous, binary, and count outcomes. Offers a unified interface for fitting and analysing panel regression models, with efficient computation for large datasets. Core numerical routines are implemented using 'Rcpp', 'RcppArmadillo', and 'RcppParallel'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CRAN.R-project.org/package=panglm}
  \item \url{https://github.com/ielbadisy/panglm}
  \item Report bugs at \url{https://github.com/ielbadisy/panglm/issues}
}

}
\author{
\strong{Maintainer}: Imad El Badisy \email{elbadisyimad@gmail.com}

}
\keyword{internal}
