% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{glance.panglm}
\alias{glance.panglm}
\title{One-row model summary of a panglm fit}
\usage{
\method{glance}{panglm}(x, ...)
}
\arguments{
\item{x}{a `"panglm"` fit}

\item{...}{unused}
}
\value{
a one-row `data.frame`
}
\description{
Implements the `generics::glance()` generic.
}
