% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panglm.R
\name{panglm}
\alias{panglm}
\title{Fit a generalized linear model for panel data}
\usage{
panglm(
  formula,
  data,
  index,
  model = c("pooling", "within", "random"),
  family = "gaussian",
  R = 21,
  maxit = 100,
  tol = 1e-10
)
}
\arguments{
\item{formula}{a model formula, e.g. `y ~ x1 + x2`}

\item{data}{a data.frame containing the variables in `formula` and `index`}

\item{index}{length-1 or length-2 character vector identifying the
individual (and optionally time) columns, e.g. `c("id", "time")`}

\item{model}{one of `"pooling"`, `"within"`, `"random"`}

\item{family}{one of `"gaussian"`, `"poisson"`, `"binomial"`, `"negbin"`,
or a family spec from [gaussian_family()] / [poisson_family()] /
[binomial_family()] / [negbin_family()]}

\item{R}{number of Gauss-Hermite quadrature nodes (random-effects
binomial models only)}

\item{maxit}{maximum IRLS/Newton iterations}

\item{tol}{convergence tolerance}
}
\value{
an object of class `"panglm"`
}
\description{
Fast pooled, fixed-effects ("within") and random-effects panel GLMs,
backed by an Rcpp/RcppArmadillo/RcppParallel numerical core. See the
package vignette for the estimators currently implemented per family.
}
