% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.panglm}
\alias{summary.panglm}
\title{Summarize a panglm fit}
\usage{
\method{summary}{panglm}(object, ...)
}
\arguments{
\item{object}{a `"panglm"` object}

\item{...}{unused}
}
\description{
Produces a classic coefficient table (Estimate, Std. Error, z value,
Pr(>|z|)) in the style of [summary.glm()].
}
